+++ crates/module-system/sov-modules-macros/tests/integration/trybuild/rpc/expose_rpc_associated_type_not_static.rs
@@ -1,14 +1,10 @@
 use jsonrpsee::core::RpcResult;
-use sov_modules_api::capabilities::mocks::MockKernel;
 use sov_modules_api::macros::{expose_rpc, rpc_gen};
 use sov_modules_api::prelude::UnwrapInfallible;
 use sov_modules_api::{
-    ApiStateAccessor, Context, DaSpec, DispatchCall, EncodeCall, Error,
-    ExecutionContext, Genesis, MessageCodec, Module, ModuleId, ModuleInfo, Spec, StateCheckpoint,
-    StateValue, TxState,
+    ApiStateAccessor, Context, DaSpec, DispatchCall, Error, Genesis, MessageCodec, Module,
+    ModuleId, ModuleInfo, Spec, StateValue, TxState,
 };
-use sov_state::ZkStorage;
-use sov_test_utils::ZkTestSpec;
 
 pub trait TestSpec: Default + std::fmt::Debug + Clone + PartialEq + Eq {
     type Data: Data;
@@ -120,34 +116,4 @@ impl TestSpec for ActualSpec {
     type Data = u32;
 }
 
-fn main() {
-    type S = ZkTestSpec;
-    type RT = Runtime<S, ActualSpec>;
-    let storage = ZkStorage::new();
-    let mut state = StateCheckpoint::new(storage, &MockKernel::<S>::default());
-    let runtime = &mut Runtime::<S, ActualSpec>::default();
-    let config = GenesisConfig::new(22);
-    let mut genesis_state = state.to_genesis_state_accessor::<RT, S>(&config);
-    runtime.genesis(&config, &mut genesis_state).unwrap();
-    let mut working_set = state.to_working_set_unmetered();
-
-    let message: u32 = 33;
-    let serialized_message =
-        <RT as EncodeCall<my_module::QueryModule<S, u32>>>::encode_call(message);
-    let module = RT::decode_call(&serialized_message, &mut working_set).unwrap();
-    let sender = <ZkTestSpec as Spec>::Address::from([11; 28]);
-    let sequencer = <ZkTestSpec as Spec>::Address::from([11; 28]);
-    let sequencer_da = <<ZkTestSpec as Spec>::Da as DaSpec>::Address::new([0; 32]);
-    let context = Context::<S>::new(
-        sender,
-        Default::default(),
-        sequencer,
-        sequencer_da,
-        1,
-        ExecutionContext::Node,
-    );
-
-    let _ = runtime
-        .dispatch_call(module, &mut working_set, &context)
-        .unwrap();
-}
+fn main() {}